chore: move generic components to base folder
