+++ packages/wallets/src/components/Base/ModalStepWrapper/ModalStepWrapper.scss
