+++ packages/wallets/src/components/Base/ModalWrapper/ModalWrapper.scss
