+++ packages/wallets/src/components/Base/ModalWrapper/ModalWrapper.tsx
@@ -1,8 +1,6 @@
 import React, { FC } from 'react';
-
-import CloseIcon from '../../public/images/ic-close-dark.svg';
-import { useModal } from '../ModalProvider';
-
+import CloseIcon from '../../../public/images/ic-close-dark.svg';
+import { useModal } from '../../ModalProvider';
 import './ModalWrapper.scss';
 
 type TProps = {
