+++ packages/wallets/src/components/Base/ProgressBar/ProgressBar.scss
