+++ packages/wallets/src/components/Base/ProgressBar/index.ts
