+++ packages/wallets/src/components/Base/Tabs/index.ts
