+++ packages/wallets/src/components/Base/WalletClipboard/WalletClipboard.scss
