+++ packages/wallets/src/components/Base/WalletClipboard/index.ts
