+++ packages/wallets/src/components/Base/index.ts
@@ -1,1 +1,8 @@
 export * from './InlineMessage';
+export * from './ModalStepWrapper';
+export * from './ModalWrapper';
+export * from './PrimaryActionButton';
+export * from './ProgressBar';
+export * from './SecondaryActionButton';
+export * from './Tabs';
+export * from './WalletClipboard';
