+++ packages/wallets/src/components/DxtradeEnterPasswordModal/DxtradeEnterPasswordModal.tsx
@@ -1,12 +1,9 @@
 import React, { useEffect, useState } from 'react';
-
 import { useActiveWalletAccount, useCreateOtherCFDAccount } from '@deriv/api';
-
 import DxTradePasswordIcon from '../../public/images/ic-dxtrade-password.svg';
+import { ModalWrapper } from '../Base';
 import { CreatePassword } from '../CreatePassword';
 import { useModal } from '../ModalProvider';
-import { ModalWrapper } from '../ModalWrapper';
-
 import './DxtradeEnterPasswordModal.scss';
 
 const DxtradeEnterPasswordModal = () => {
