+++ packages/wallets/src/components/JurisdictionModal/JurisdictionModal.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import classNames from 'classnames';
+import { ModalStepWrapper } from '../Base';
 import { useModal } from '../ModalProvider';
-import { ModalStepWrapper } from '../ModalStepWrapper';
 import { MT5PasswordModal } from '../MT5PasswordModal';
 import JurisdictionCard from './JurisdictionCard';
 import './JurisdictionModal.scss';
