+++ packages/wallets/src/components/index.ts
@@ -9,13 +9,9 @@ export * from './Loader';
 export * from './MT5AccountTypeModal';
 export * from './MT5PasswordModal';
 export * from './OptionsAndMultipliersListing';
-export * from './PrimaryActionButton';
-export * from './ProgressBar';
-export * from './SecondaryActionButton';
 export * from './TradingAccountCard';
 export * from './WalletCard';
 export * from './WalletCardIcon';
-export * from './WalletClipboard';
 export * from './WalletListCard';
 export * from './WalletListCardActions';
 export * from './WalletListCardBadge';
