+++ packages/wallets/src/features/cashier/screens/WalletDepositScreens/WalletDepositCryptoAddress/WalletDepositCryptoAddress.tsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 import QRCode from 'qrcode.react';
 import { useAuthorize, useDepositCryptoAddress } from '@deriv/api';
+import { WalletClipboard } from '../../../../../components/Base';
 import { Loader } from '../../../../../components/Loader';
-import { WalletClipboard } from '../../../../../components/WalletClipboard';
 import useDevice from '../../../../../hooks/useDevice';
 import './WalletDepositCryptoAddress.scss';
